simplify agent implementation
