--- agent/internal/agentopt/README.go
@@ -1,4 +1,0 @@
-// Package agentopt contains internal option plumbing.
-//
-// See README.md for rationale and package design notes.
-package agentopt
