+++ agent/middleware/structuredoutput/structuredoutput_test.go
@@ -9,9 +9,8 @@ import (
 	"iter"
 	"testing"
 
-	"github.com/microsoft/agent-framework-go/agent/internal/agentopt"
-	"github.com/microsoft/agent-framework-go/agent/internal/middleware"
-	"github.com/microsoft/agent-framework-go/agent/internal/middleware/structuredoutput"
+	"github.com/microsoft/agent-framework-go/agent"
+	"github.com/microsoft/agent-framework-go/agent/middleware/structuredoutput"
 	"github.com/microsoft/agent-framework-go/format"
 	"github.com/microsoft/agent-framework-go/message"
 )
@@ -27,7 +26,7 @@ func (t testFormat) Kind() string {
 func TestStructuredOutput_NoStructuredOutputOption(t *testing.T) {
 	// Test that middleware passes through when no structured output option is provided
 	baseCalled := false
-	baseFunc := func(ctx context.Context, messages []*message.Message, options ...agentopt.Option) iter.Seq2[*message.ResponseUpdate, error] {
+	baseFunc := func(ctx context.Context, messages []*message.Message, options ...agent.Option) iter.Seq2[*message.ResponseUpdate, error] {
 		baseCalled = true
 		return func(yield func(*message.ResponseUpdate, error) bool) {
 			yield(&message.ResponseUpdate{
@@ -51,7 +50,7 @@ func TestStructuredOutput_NoStructuredOutputOption(t *testing.T) {
 	ctx := context.Background()
 
 	var updates []*message.ResponseUpdate
-	seq := middleware.RunChain(ctx, baseFunc, []middleware.Middleware{mw}, []*message.Message{})
+	seq := mw.Run(baseFunc, ctx, []*message.Message{})
 	for update, err := range seq {
 		if err != nil {
 			t.Fatalf("unexpected error: %v", err)
@@ -75,7 +74,7 @@ func TestStructuredOutput_NoStructuredOutputOption(t *testing.T) {
 func TestStructuredOutput_NilStructuredOutputOption(t *testing.T) {
 	// Test that middleware passes through when structured output option is nil
 	baseCalled := false
-	baseFunc := func(ctx context.Context, messages []*message.Message, options ...agentopt.Option) iter.Seq2[*message.ResponseUpdate, error] {
+	baseFunc := func(ctx context.Context, messages []*message.Message, options ...agent.Option) iter.Seq2[*message.ResponseUpdate, error] {
 		baseCalled = true
 		return func(yield func(*message.ResponseUpdate, error) bool) {
 			yield(&message.ResponseUpdate{
@@ -99,7 +98,7 @@ func TestStructuredOutput_NilStructuredOutputOption(t *testing.T) {
 	ctx := context.Background()
 
 	var updates []*message.ResponseUpdate
-	seq := middleware.RunChain(ctx, baseFunc, []middleware.Middleware{mw}, []*message.Message{}, agentopt.WithStructuredOutput(nil))
+	seq := mw.Run(baseFunc, ctx, []*message.Message{}, agent.WithStructuredOutput(nil))
 	for update, err := range seq {
 		if err != nil {
 			t.Fatalf("unexpected error: %v", err)
@@ -118,7 +117,7 @@ func TestStructuredOutput_NilStructuredOutputOption(t *testing.T) {
 
 func TestStructuredOutput_MissingFormat(t *testing.T) {
 	// Test that middleware returns error when Format is nil
-	baseFunc := func(ctx context.Context, messages []*message.Message, options ...agentopt.Option) iter.Seq2[*message.ResponseUpdate, error] {
+	baseFunc := func(ctx context.Context, messages []*message.Message, options ...agent.Option) iter.Seq2[*message.ResponseUpdate, error] {
 		return func(yield func(*message.ResponseUpdate, error) bool) {
 			yield(&message.ResponseUpdate{
 				Contents: message.Contents{
@@ -139,7 +138,7 @@ func TestStructuredOutput_MissingFormat(t *testing.T) {
 	ctx := context.Background()
 	output := &struct{ Name string }{}
 
-	seq := middleware.RunChain(ctx, baseFunc, []middleware.Middleware{mw}, []*message.Message{}, agentopt.WithStructuredOutput(output))
+	seq := mw.Run(baseFunc, ctx, []*message.Message{}, agent.WithStructuredOutput(output))
 	for _, err := range seq {
 		if err == nil {
 			t.Fatal("expected error for missing Format")
@@ -154,7 +153,7 @@ func TestStructuredOutput_MissingFormat(t *testing.T) {
 
 func TestStructuredOutput_MissingUnmarshal(t *testing.T) {
 	// Test that middleware returns error when Unmarshal is nil
-	baseFunc := func(ctx context.Context, messages []*message.Message, options ...agentopt.Option) iter.Seq2[*message.ResponseUpdate, error] {
+	baseFunc := func(ctx context.Context, messages []*message.Message, options ...agent.Option) iter.Seq2[*message.ResponseUpdate, error] {
 		return func(yield func(*message.ResponseUpdate, error) bool) {
 			yield(&message.ResponseUpdate{
 				Contents: message.Contents{
@@ -175,7 +174,7 @@ func TestStructuredOutput_MissingUnmarshal(t *testing.T) {
 	ctx := context.Background()
 	output := &struct{ Name string }{}
 
-	seq := middleware.RunChain(ctx, baseFunc, []middleware.Middleware{mw}, []*message.Message{}, agentopt.WithStructuredOutput(output))
+	seq := mw.Run(baseFunc, ctx, []*message.Message{}, agent.WithStructuredOutput(output))
 	for _, err := range seq {
 		if err == nil {
 			t.Fatal("expected error for missing Unmarshal")
@@ -190,7 +189,7 @@ func TestStructuredOutput_MissingUnmarshal(t *testing.T) {
 
 func TestStructuredOutput_FormatError(t *testing.T) {
 	// Test that middleware propagates Format errors
-	baseFunc := func(ctx context.Context, messages []*message.Message, options ...agentopt.Option) iter.Seq2[*message.ResponseUpdate, error] {
+	baseFunc := func(ctx context.Context, messages []*message.Message, options ...agent.Option) iter.Seq2[*message.ResponseUpdate, error] {
 		return func(yield func(*message.ResponseUpdate, error) bool) {
 			yield(&message.ResponseUpdate{
 				Contents: message.Contents{
@@ -214,7 +213,7 @@ func TestStructuredOutput_FormatError(t *testing.T) {
 	ctx := context.Background()
 	output := &struct{ Name string }{}
 
-	seq := middleware.RunChain(ctx, baseFunc, []middleware.Middleware{mw}, []*message.Message{}, agentopt.WithStructuredOutput(output))
+	seq := mw.Run(baseFunc, ctx, []*message.Message{}, agent.WithStructuredOutput(output))
 	for _, err := range seq {
 		if err == nil {
 			t.Fatal("expected error for Format failure")
@@ -229,10 +228,10 @@ func TestStructuredOutput_FormatError(t *testing.T) {
 
 func TestStructuredOutput_SuccessfulUnmarshal(t *testing.T) {
 	// Test successful structured output parsing
-	baseFunc := func(ctx context.Context, messages []*message.Message, options ...agentopt.Option) iter.Seq2[*message.ResponseUpdate, error] {
+	baseFunc := func(ctx context.Context, messages []*message.Message, options ...agent.Option) iter.Seq2[*message.ResponseUpdate, error] {
 		return func(yield func(*message.ResponseUpdate, error) bool) {
 			// Verify that ResponseFormat option was added
-			v, ok := agentopt.GetOption(options, agentopt.WithResponseFormat)
+			v, ok := agent.GetOption(options, agent.WithResponseFormat)
 			if !ok {
 				yield(nil, errors.New("ResponseFormat option not found"))
 				return
@@ -281,7 +280,7 @@ func TestStructuredOutput_SuccessfulUnmarshal(t *testing.T) {
 		Age  int    `json:"age"`
 	}{}
 
-	seq := middleware.RunChain(ctx, baseFunc, []middleware.Middleware{mw}, []*message.Message{}, agentopt.WithStructuredOutput(output))
+	seq := mw.Run(baseFunc, ctx, []*message.Message{}, agent.WithStructuredOutput(output))
 	for _, err := range seq {
 		if err != nil {
 			t.Fatalf("unexpected error: %v", err)
@@ -298,7 +297,7 @@ func TestStructuredOutput_SuccessfulUnmarshal(t *testing.T) {
 
 func TestStructuredOutput_UnmarshalError(t *testing.T) {
 	// Test that middleware propagates Unmarshal errors
-	baseFunc := func(ctx context.Context, messages []*message.Message, options ...agentopt.Option) iter.Seq2[*message.ResponseUpdate, error] {
+	baseFunc := func(ctx context.Context, messages []*message.Message, options ...agent.Option) iter.Seq2[*message.ResponseUpdate, error] {
 		return func(yield func(*message.ResponseUpdate, error) bool) {
 			yield(&message.ResponseUpdate{
 				Contents: message.Contents{
@@ -322,7 +321,7 @@ func TestStructuredOutput_UnmarshalError(t *testing.T) {
 	ctx := context.Background()
 	output := &struct{ Name string }{}
 
-	seq := middleware.RunChain(ctx, baseFunc, []middleware.Middleware{mw}, []*message.Message{}, agentopt.WithStructuredOutput(output))
+	seq := mw.Run(baseFunc, ctx, []*message.Message{}, agent.WithStructuredOutput(output))
 	for _, err := range seq {
 		if err == nil {
 			t.Fatal("expected error for Unmarshal failure")
@@ -338,7 +337,7 @@ func TestStructuredOutput_UnmarshalError(t *testing.T) {
 func TestStructuredOutput_BaseErrorPropagation(t *testing.T) {
 	// Test that errors from the base function are propagated
 	expectedErr := errors.New("base function error")
-	baseFunc := func(ctx context.Context, messages []*message.Message, options ...agentopt.Option) iter.Seq2[*message.ResponseUpdate, error] {
+	baseFunc := func(ctx context.Context, messages []*message.Message, options ...agent.Option) iter.Seq2[*message.ResponseUpdate, error] {
 		return func(yield func(*message.ResponseUpdate, error) bool) {
 			yield(nil, expectedErr)
 		}
@@ -357,7 +356,7 @@ func TestStructuredOutput_BaseErrorPropagation(t *testing.T) {
 	ctx := context.Background()
 	output := &struct{ Name string }{}
 
-	seq := middleware.RunChain(ctx, baseFunc, []middleware.Middleware{mw}, []*message.Message{}, agentopt.WithStructuredOutput(output))
+	seq := mw.Run(baseFunc, ctx, []*message.Message{}, agent.WithStructuredOutput(output))
 	for _, err := range seq {
 		if err == nil {
 			t.Fatal("expected error from base function")
@@ -383,7 +382,7 @@ func TestStructuredOutput_ComplexStructure(t *testing.T) {
 		Tags    []string `json:"tags"`
 	}
 
-	baseFunc := func(ctx context.Context, messages []*message.Message, options ...agentopt.Option) iter.Seq2[*message.ResponseUpdate, error] {
+	baseFunc := func(ctx context.Context, messages []*message.Message, options ...agent.Option) iter.Seq2[*message.ResponseUpdate, error] {
 		return func(yield func(*message.ResponseUpdate, error) bool) {
 			jsonStr := `{"name":"Bob","age":25,"address":{"street":"123 Main St","city":"Springfield"},"tags":["developer","golang"]}`
 			yield(&message.ResponseUpdate{
@@ -407,7 +406,7 @@ func TestStructuredOutput_ComplexStructure(t *testing.T) {
 	ctx := context.Background()
 	output := &Person{}
 
-	seq := middleware.RunChain(ctx, baseFunc, []middleware.Middleware{mw}, []*message.Message{}, agentopt.WithStructuredOutput(output))
+	seq := mw.Run(baseFunc, ctx, []*message.Message{}, agent.WithStructuredOutput(output))
 	for _, err := range seq {
 		if err != nil {
 			t.Fatalf("unexpected error: %v", err)
@@ -436,7 +435,7 @@ func TestStructuredOutput_ComplexStructure(t *testing.T) {
 
 func TestStructuredOutput_EmptyResponse(t *testing.T) {
 	// Test with an empty response
-	baseFunc := func(ctx context.Context, messages []*message.Message, options ...agentopt.Option) iter.Seq2[*message.ResponseUpdate, error] {
+	baseFunc := func(ctx context.Context, messages []*message.Message, options ...agent.Option) iter.Seq2[*message.ResponseUpdate, error] {
 		return func(yield func(*message.ResponseUpdate, error) bool) {
 			// Empty response
 		}
@@ -458,7 +457,7 @@ func TestStructuredOutput_EmptyResponse(t *testing.T) {
 	ctx := context.Background()
 	output := &struct{ Name string }{}
 
-	seq := middleware.RunChain(ctx, baseFunc, []middleware.Middleware{mw}, []*message.Message{}, agentopt.WithStructuredOutput(output))
+	seq := mw.Run(baseFunc, ctx, []*message.Message{}, agent.WithStructuredOutput(output))
 	for _, err := range seq {
 		if err == nil {
 			t.Fatal("expected error for empty response")
@@ -473,7 +472,7 @@ func TestStructuredOutput_EmptyResponse(t *testing.T) {
 
 func TestStructuredOutput_MultipleContentTypes(t *testing.T) {
 	// Test that only text content is collected for unmarshaling
-	baseFunc := func(ctx context.Context, messages []*message.Message, options ...agentopt.Option) iter.Seq2[*message.ResponseUpdate, error] {
+	baseFunc := func(ctx context.Context, messages []*message.Message, options ...agent.Option) iter.Seq2[*message.ResponseUpdate, error] {
 		return func(yield func(*message.ResponseUpdate, error) bool) {
 			if !yield(&message.ResponseUpdate{
 				Contents: message.Contents{
@@ -516,7 +515,7 @@ func TestStructuredOutput_MultipleContentTypes(t *testing.T) {
 		Name string `json:"name"`
 	}{}
 
-	seq := middleware.RunChain(ctx, baseFunc, []middleware.Middleware{mw}, []*message.Message{}, agentopt.WithStructuredOutput(output))
+	seq := mw.Run(baseFunc, ctx, []*message.Message{}, agent.WithStructuredOutput(output))
 	for _, err := range seq {
 		if err != nil {
 			t.Fatalf("unexpected error: %v", err)
