+++ agent/provider/openaiagent/chat.go
@@ -9,9 +9,12 @@ import (
 	"fmt"
 	"iter"
 	"reflect"
+	"slices"
 	"time"
 
 	"github.com/microsoft/agent-framework-go/agent"
+	"github.com/microsoft/agent-framework-go/agent/middleware/autocall"
+	"github.com/microsoft/agent-framework-go/agent/middleware/structuredoutput"
 	"github.com/microsoft/agent-framework-go/format"
 	"github.com/microsoft/agent-framework-go/format/jsonformat"
 	"github.com/microsoft/agent-framework-go/message"
@@ -50,10 +53,19 @@ func NewChatCompletions(oclient openai.Client, config Config) *agent.Agent {
 		client: oclient,
 		config: config,
 	}
+	config.Config.Middlewares = slices.Clone(config.Config.Middlewares)
+	if !config.Config.DisableFuncAutoCall {
+		config.Config.Middlewares = append(config.Config.Middlewares, autocall.New(autocall.Config{
+			Logger:           config.Config.Logger,
+			LogSensitiveData: config.Config.LogSensitiveData,
+		}))
+	}
+	config.Config.Middlewares = append(config.Config.Middlewares, structuredoutput.New(structuredoutput.Config{
+		Format:    c.formatOf,
+		Unmarshal: c.unmarshal,
+	}))
 	return agent.New(agent.ProviderConfig{
 		ProviderName: "openai",
-		FormatOfFn:   c.formatOf,
-		UnmarshalFn:  c.unmarshal,
 		Run:          c.run,
 	}, config.Config)
 }
